refactor(workshop): extract widget state architecture
