+++ .todo/tech-debt/README.md
@@ -65,7 +65,8 @@ elsewhere.
 | Low | [Build command gate semantics](2026-06-29-build-command-gate-semantics.md) | Deferred |
 | Low | [Workshop bounded turn packer](2026-07-14-workshop-bounded-turn-packer.md) | Deferred |
 | Medium | [Persona prompt/validator contract drift](2026-07-26-persona-prompt-validator-contract-drift.md) | Open |
-| Medium | [Workshop widget-local persisted codecs](2026-07-30-workshop-widget-local-codecs.md) | Deferred |
+| Medium | [Workshop widget-local persisted codecs](2026-07-30-workshop-widget-local-codecs.md) | In Progress |
+| Low | [Widget-config counter integrity validation](2026-07-31-widget-config-counter-integrity.md) | Identified |
 
 ## Review Guidance
